Add function insert(value)
This function calls another function that checks up to 3
conditions before inserting the new value/node:
1.If the root is null it creates a new node with the given value.
2.If the root is not null, it checks if the root value is equal to the
given value, in which case it returns the root without changes.
3.If the given value is different than the root value, it checks if
it is greater or lower than it.
After repeating this process recursively on the sub nodes,
it creates the new node with the given value and returns the updated
root node.