Add function delete(value)
This function searches for the given value recursively and removes it.
If the value is not found anywhere it returns the same unaltered root
node.
The hardest part is removing a parent node with both left and right
children, in which case the function will find the leftmost node
greater than the parent node and replace the parent node with it while
maintaining the binary search tree intact.