Add function height(node)
This function returns the number of edges in the longest path from
the given node to a leaf node.
If the given node has both left and right nodes, it will compare the
distance from their leaf nodes and return the greatest number.
if the node doesn't exist it simply returns undefined.